split storage2 module
