+++ tkcore/cpp_lib/CMakeLists.txt
@@ -16,12 +16,12 @@ add_library(lib_tkcore
 
         # storage 2
 
-        ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt2.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt2_schema.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/salt_text/salt_base.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/main/storage2/split_password.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/main/storage2/storage.cpp
-        ${CMAKE_CURRENT_LIST_DIR}/main/storage2/storage_structure.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/storage2/salt_text/salt2.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/storage2/salt_text/salt2_schema.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/storage2/salt_text/salt_base.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/storage2/split_password.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/storage2/storage.cpp
+        ${CMAKE_CURRENT_LIST_DIR}/storage2/storage_structure.cpp
 )
 
 set_target_properties(lib_tkcore PROPERTIES INTERFACE_INCLUDE_DIRECTORIES ${CMAKE_CURRENT_LIST_DIR})
