+++ tkcore/cpp_lib/storage1/key_storage_v1.cpp
@@ -3,7 +3,7 @@
 //
 
 #include "core/key_core.h"
-#include "key_storage_v1.h"
+#include "storage.h"
 #include "core/key_errors.h"
 #include "pass_spliter_v1.h"
 #include "core/common.h"
