+++ tkcore/cpp_lib/storage1/storage.h
@@ -103,4 +103,4 @@ namespace thekey_v1 {
 
 }
 
-#endif //THEKEY_KEY_STORAGE_V1_H
+#endif //THEKEY_STORAGE_H
