+++ tkcore/cpp_lib/storage2/salt_text/salt2.cpp
