+++ tkcore/cpp_lib/storage2/salt_text/salt2_schema.cpp
