+++ tkcore/cpp_lib/storage2/storage.cpp
