+++ tkcore/cpp_lib/storage2/storage_structure.cpp
